refactor: generalize and DRY up lambda-api gateway integration (#1)
* refactor: move utility functions & constructs to utils/
* fix: lambda build directory path
* chore: target ES2020
* chore: pin specific package versions
* refactor: DRY up lambda integration and method adding
* fix: create unique versioning for lambdas to address reload warning
