+++ package.json
@@ -17,15 +17,15 @@
   },
   "license": "Apache-2.0",
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.130",
-    "@types/node": "*",
-    "aws-cdk": "*",
-    "aws-cdk-local": "^2.18.0",
-    "esbuild": "^0.19.10",
-    "typescript": "~5.1.6"
+    "@types/aws-lambda": "8.10.152",
+    "@types/node": "24.3.0",
+    "aws-cdk": "2.1025.0",
+    "aws-cdk-local": "3.0.1",
+    "esbuild": "0.25.9",
+    "typescript": "5.9.2"
   },
   "dependencies": {
-    "aws-cdk-lib": "^2.0.0",
-    "constructs": "^10.0.0"
+    "aws-cdk-lib": "2.211.0",
+    "constructs": "10.4.2"
   }
 }
