+++ tsconfig.json
@@ -1,12 +1,8 @@
 {
   "compilerOptions": {
-    "target": "ES2018",
+    "target": "ES2020",
     "module": "commonjs",
-    "lib": [
-      "es2016",
-      "es2017.object",
-      "es2017.string"
-    ],
+    "lib": ["es2016", "es2017.object", "es2017.string"],
     "strict": true,
     "noImplicitAny": true,
     "strictNullChecks": true,
@@ -21,7 +17,5 @@
     "experimentalDecorators": true,
     "strictPropertyInitialization": false
   },
-  "include": [
-    "*.ts"
-  ]
+  "include": ["*.ts"]
 }
