+++ utils/addCorsOptions.ts
@@ -1,0 +1,47 @@
+import {
+  IResource,
+  MockIntegration,
+  PassthroughBehavior,
+} from "aws-cdk-lib/aws-apigateway";
+
+export function addCorsOptions(apiResource: IResource) {
+  apiResource.addMethod(
+    "OPTIONS",
+    new MockIntegration({
+      // In case you want to use binary media types, uncomment the following line
+      // contentHandling: ContentHandling.CONVERT_TO_TEXT,
+      integrationResponses: [
+        {
+          statusCode: "200",
+          responseParameters: {
+            "method.response.header.Access-Control-Allow-Headers":
+              "'Content-Type,X-Amz-Date,Authorization,X-Api-Key,X-Amz-Security-Token,X-Amz-User-Agent'",
+            "method.response.header.Access-Control-Allow-Origin": "'*'",
+            "method.response.header.Access-Control-Allow-Credentials":
+              "'false'",
+            "method.response.header.Access-Control-Allow-Methods":
+              "'OPTIONS,GET,PUT,POST,DELETE'",
+          },
+        },
+      ],
+      // In case you want to use binary media types, comment out the following line
+      passthroughBehavior: PassthroughBehavior.NEVER,
+      requestTemplates: {
+        "application/json": '{"statusCode": 200}',
+      },
+    }),
+    {
+      methodResponses: [
+        {
+          statusCode: "200",
+          responseParameters: {
+            "method.response.header.Access-Control-Allow-Headers": true,
+            "method.response.header.Access-Control-Allow-Methods": true,
+            "method.response.header.Access-Control-Allow-Credentials": true,
+            "method.response.header.Access-Control-Allow-Origin": true,
+          },
+        },
+      ],
+    }
+  );
+}
